Edit Achievement Working Properly

+++ Server/src/main/java/com/abs/errp/achievement/AchievementRepository.java
@@ -6,12 +6,9 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 
 import com.abs.errp.entity.Achievement;
 
-public interface AchievementRepository extends PagingAndSortingRepository<Achievement, Long>  {
-	
-	
-
-	List<Achievement> findByEmployeeId(int employeeId);
-
+public interface AchievementRepository extends PagingAndSortingRepository<Achievement, Integer>  {
+	  List<Achievement> findByEmployeeId(int employeeId);
+     
 	
 
 }
