+++ Server/src/main/java/com/abs/errp/achievement/AchievementService.java
@@ -11,7 +11,7 @@ public interface AchievementService {
 	Achievement saveAchievement(Achievement achievement);
 	List<Achievement> findPaginated(int pageNo,int pageSize);
 	// edit 
-	Achievement getAchievementById(long id);
+	Achievement getAchievementById(int id);
 	//update the achievement after click the update button 
-	Achievement updateAchievement(Long id, Achievement achievement);
+	Achievement updateAchievement(int id, Achievement achievement);
 }
